+++ packages/f-header/src/components/Header.vue
@@ -1,14 +1,14 @@
 <template>
     <header
         :data-theme="theme"
-        :class="['c-header', { 'c-header--transparent c-header--gradient': isTransparent }]">
+        :class="['c-header', { 'c-header--transparent c-header--gradient': showTransparentHeader }]">
         <skip-to-main
             :text="copy.skipToMainContentText"
-            :transparent-bg="isTransparent" />
+            :transparent-bg="showTransparentHeader" />
         <div class="c-header-container">
             <logo
                 :theme="theme"
-                :is-transparent="isTransparent"
+                :is-transparent="showTransparentHeader"
                 :company-name="copy.companyName"
                 :logo-gtm-label="copy.logo.gtm" />
             <navigation
@@ -20,7 +20,8 @@
                 :delivery-enquiry="copy.deliveryEnquiry"
                 :show-delivery-enquiry="showDeliveryEnquiryWithContent"
                 :user-info-prop="userInfoProp"
-                :just-log="justLog" />
+                :just-log="justLog"
+                @onMobileNavToggle="mobileNavToggled" />
         </div>
     </header>
 </template>
@@ -42,12 +43,10 @@ export default {
     props: {
         locale: {
             type: String,
-            required: false,
             default: ''
         },
         isTransparent: {
             type: Boolean,
-            required: false,
             default: false
         },
         showDeliveryEnquiry: {
@@ -56,28 +55,38 @@ export default {
         },
         justLog: {
             type: Object,
-            default: () => ({}),
-            required: false
+            default: () => ({})
         },
         userInfoProp: {
             type: [Object, Boolean],
-            default: false,
-            required: false
+            default: false
         }
     },
     data () {
         const locale = sharedServices.getLocale(tenantConfigs, this.locale, this.$118n);
         const localeConfig = tenantConfigs[locale];
-        const theme = sharedServices.getTheme(locale);
+        const theme = sharedService.getTheme(locale);
+        const mobileNavIsOpen = false;
 
         return {
             copy: { ...localeConfig },
-            theme
+            theme,
+            mobileNavIsOpen
         };
     },
     computed: {
         showDeliveryEnquiryWithContent () {
             return this.copy.deliveryEnquiry && this.showDeliveryEnquiry;
+        },
+        showTransparentHeader () {
+            return this.isTransparent && !this.mobileNavIsOpen;
+        }
+    },
+    methods: {
+        // This method emits `navIsOpen` state from the navigation component
+        // to be able to deside when to show transparent header styles on mobile view
+        mobileNavToggled (value) {
+            this.mobileNavIsOpen = value;
         }
     }
 };