+++ packages/f-header/CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v2.0.0-beta-11
+------------------------------
+*September 25, 2019*
+
+### Added
+- `no-js` css for mobile hamburger menu
+
+### Changed
+- Styling for mobile open menu state header when transparent
+
+
 v2.0.0-beta-10
 ------------------------------
 *September 20, 2019*